1畅游寰宇源程序包/requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.7.1
bs4==0.0.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
jdcal==1.4
jieba==0.39
lxml==4.3.0
mysqlclient==1.4.2.post1
numpy==1.16.2
odfpy==1.4.0
openpyxl==2.6.1
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1
six==1.12.0
soupsieve==1.7.3
tablib==0.13.0
tqdm==4.31.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
